Updating the BoundedRange to use the size of the system architecture.

+++ Sources/VimKitShaders/include/ShaderTypes.h
@@ -23,9 +23,9 @@ typedef NSInteger EnumBackingType;
 
 typedef struct {
     // The lower bounds of the range
-    uint32_t lowerBound;
+    size_t lowerBound;
     // The upper bounds of the range
-    uint32_t upperBound;
+    size_t upperBound;
 } BoundedRange;
 
 typedef struct {
